+++ lmk/__main__.py
@@ -1,4 +1,4 @@
-from lmk.cli.main import cli
+from lmk.cli import cli
 
 
 if __name__ == "__main__":